Move RecipeValidator, fix comments typos
